+++ .rtx.toml
@@ -1,0 +1,6 @@
+[env]
+NEXT_PUBLIC_ERDDAP_SERVICE = "https://buoybarn.neracoos.org"
+
+[tools]
+node = "20"
+yarn = "1.22.19"
