+++ .storybook/preview.tsx
@@ -1,0 +1,18 @@
+import type { Preview } from "@storybook/react"
+import { MemoryRouter } from "react-router-dom"
+
+import "../src/index.scss"
+
+const preview: Preview = {
+  parameters: {
+    actions: { argTypesRegex: "^on[A-Z].*" },
+    controls: {
+      matchers: {
+        color: /(background|color)$/i,
+        date: /Date$/,
+      },
+    },
+  },
+}
+
+export default preview
