+++ package.json
@@ -54,16 +54,20 @@
     "docs": "typedoc",
     "test-with-coverage": "jest --verbose --coverage && cat ./coverage/lcov.info",
     "types": "tsc",
-    "storybook": "storybook dev -p 9009 -s public",
-    "build-storybook": "storybook build -c .storybook -o storybook-public"
+    "storybook": "storybook dev -p 6006",
+    "build-storybook": "storybook build"
   },
   "devDependencies": {
-    "@storybook/addon-a11y": "^7.3.2",
-    "@storybook/addon-essentials": "^7.3.2",
-    "@storybook/addon-storyshots": "^7.3.2",
-    "@storybook/preset-create-react-app": "^7.3.2",
-    "@storybook/react": "^7.3.2",
-    "@storybook/react-webpack5": "^7.3.2",
+    "@storybook/addon-a11y": "^7.4.6",
+    "@storybook/addon-essentials": "^7.4.6",
+    "@storybook/addon-interactions": "^7.4.6",
+    "@storybook/addon-links": "^7.4.6",
+    "@storybook/addon-onboarding": "^1.0.8",
+    "@storybook/addon-storyshots": "^7.4.6",
+    "@storybook/blocks": "^7.4.6",
+    "@storybook/nextjs": "^7.4.6",
+    "@storybook/react": "^7.4.6",
+    "@storybook/testing-library": "^0.2.2",
     "@testing-library/dom": "^9.3.1",
     "@testing-library/jest-dom": "^6.1.2",
     "@testing-library/react": "^14.0.0",
@@ -73,12 +77,13 @@
     "@types/react": "18.2.21",
     "@types/react-dom": "18.2.7",
     "@types/react-test-renderer": "18.0.0",
+    "eslint-plugin-storybook": "^0.6.15",
     "jest-environment-jsdom": "^29.6.4",
     "jest-fetch-mock": "3.0.3",
     "react-test-renderer": "18.2.0",
     "redux-devtools-extension": "2.13.9",
     "redux-mock-store": "1.5.4",
-    "storybook": "^7.3.2",
+    "storybook": "^7.4.6",
     "typescript": "^4.8.4",
     "wait-on": "7.0.1",
     "webpack": "5"
@@ -104,7 +109,8 @@
   "eslintConfig": {
     "extends": [
       "react-app",
-      "react-app/jest"
+      "react-app/jest",
+      "plugin:storybook/recommended"
     ],
     "ignorePatterns": [
       "**/*.stories.tsx"
