+++ src/components/NavBar/index.stories.tsx
@@ -1,10 +1,23 @@
+import type { Meta, StoryObj } from '@storybook/react';
 import * as React from "react"
+import { MemoryRouter } from "react-router-dom"
 
 import NeracoosNavBar from "./index"
 
-export default {
+const meta: Meta<typeof NeracoosNavBar> = {
   component: NeracoosNavBar,
   title: "Components/NavBar",
+  decorators: [
+    (Story) => (
+      <MemoryRouter>
+        <Story />
+      </MemoryRouter>
+    )
+  ]
 }
 
-export const navBar = () => <NeracoosNavBar />
+export default meta
+
+type Story = StoryObj<typeof NeracoosNavBar>
+
+export const NavBar: Story = {}
